Fix versioning for editable installs
This was missed when we went to the src layout
